+++ PlatformerBase/Assets/Scripts/UI/Heartbeat.cs
@@ -137,7 +137,7 @@ public class Heartbeat : Global {
         if (virus)
         {
             IHealthObject player = Player.GetComponent<IHealthObject>();
-            virus.virusMat.SetFloat("_Cutoff", Mathf.Lerp(virus.virusMat.GetFloat("_Cutoff"), 1 - (player.Health / (float)player.MaxHealth), 0.2f));
+            virus.virusMat.SetFloat("_Cutoff", Mathf.Lerp(virus.virusMat.GetFloat("_Cutoff"), 1 - (player.Health / player.MaxHealth), 0.2f));
         }
        
     }
